// src/app/models/producto.dmodificar.model.ts
export interface ProductosAModificar {
  COD_PRODUCTO: string;
  NOMBRE_PRODUCTO: string;
  COD_CATEGORIA: string;
  COD_SUB_CATEGORIA: string;
  COSTO_TOTAL_PRODUCTO: number;
  FECHA_VENCIMIENTO: string;
  ID_PROVEEDOR: string;
  DESCRIPCION: string;
  STOCK: number;
  ENUSO: number;
  ESTADO_CRITICO: number;
  ID_PRESENTACION: number;
  ID_UNIDADES_PRESENTACION: number;
  UNIDADES: number;
  MONEDA: string;
}
/// Este modelo es cuando se van a modificar los datos